python-backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
PyPDF2==3.0.1
reportlab==3.6.1
Pillow==8.3.2
gunicorn==20.1.0
